Move main rebuild_quotas logic to plugins/quotas
